+++ docs/source/conf.py
@@ -26,7 +26,7 @@ copyright = '2022, Moritz Layer, Johanna Senk, Simon Essink, Alexander van Meege
 author = 'Moritz Layer, Johanna Senk, Simon Essink, Alexander van Meegen, Hannah Bos, Moritz Helias'
 
 # The full version, including alpha/beta/rc tags
-release = '1.0.2'
+release = '1.1.0'
 
 
 # -- General configuration ---------------------------------------------------